+++ src/components/Calendar.vue
@@ -12,8 +12,8 @@ import {
   getNextPage,
   getPageBetweenPages,
   getFirstValidPage,
+  getPageForDate,
 } from '../utils/helpers';
-import '../styles/lib.sass';
 
 export default {
   mixins: [mergeListeners],
@@ -24,13 +24,23 @@ export default {
           ...this.$attrs,
           position,
           page: position < 2 ? this.fromPage_ : this.toPage_,
-          minPage: position < 2 ? this.minPage : this.minToPage,
-          maxPage: position < 2 ? this.maxFromPage : this.maxPage,
+          minPage: position < 2 ? this.minPage_ : this.minToPage,
+          maxPage: position < 2 ? this.maxFromPage : this.maxPage_,
+          hideRightButton:
+            !this.showLinkedButtons &&
+            position === 1 &&
+            this.isLinked &&
+            !this.isVertical,
+          hideLeftButton:
+            !this.showLinkedButtons &&
+            position === 2 &&
+            this.isLinked &&
+            !this.isVertical,
+          paneWidth: this.paneWidth,
           styles: this.themeStyles_,
           attributes: this.attributes_,
         },
         on: this.mergeListeners({
-          titleclick: this.titleClick,
           'update:page': val => {
             if (position < 2) this.fromPage_ = val;
             else this.toPage_ = val;
@@ -44,12 +54,22 @@ export default {
       {
         class: {
           'c-pane-container': true,
-          'is-double-paned': this.isDoublePaned_,
+          'is-vertical': this.isVertical,
           'is-expanded': this.isExpanded,
         },
         style: this.wrapperStyle,
+        ref: 'root',
       },
-      [getPaneComponent(this.isDoublePaned_ ? 1 : 0), this.isDoublePaned_ && getPaneComponent(2)],
+      [
+        getPaneComponent(this.isDoublePaned_ ? 1 : 0),
+        ...(this.isDoublePaned_ && [
+          h('div', {
+            class: 'c-pane-div',
+            style: this.dividerStyle,
+          }),
+          getPaneComponent(2),
+        ]),
+      ],
     );
   },
   name: 'VCalendar',
@@ -57,36 +77,57 @@ export default {
     CalendarPane,
   },
   props: {
+    minDate: Date,
+    maxDate: Date,
     minPage: Object,
     maxPage: Object,
     fromPage: Object,
     toPage: Object,
+    showLinkedButtons: {
+      type: Boolean,
+      default: () => defaults.showLinkedButtons,
+    },
     isDoublePaned: Boolean,
+    isLinked: Boolean,
+    isVertical: Boolean,
     isExpanded: Boolean,
-    paneWidth: { type: Number, default: defaults.paneWidth },
+    paneWidth: { type: Number, default: () => defaults.paneWidth },
     themeStyles: Object,
     attributes: Array,
   },
   data() {
     return {
-      windowWidth: 0,
+      isConstrained: true,
       fromPage_: null,
       toPage_: null,
     };
   },
   computed: {
     isDoublePaned_() {
-      return this.isDoublePaned && this.windowWidth >= 2 * this.paneWidth + 30;
+      return this.isDoublePaned && (this.isVertical || !this.isConstrained);
     },
-    paneCentered() {
-      return this.isDoublePaned && !this.isDoublePaned_;
+    minPage_() {
+      return (
+        this.minPage || (this.minDate && getPageForDate(this.minDate)) || null
+      );
+    },
+    rightButtonHidden() {
+      return this.position === 1 && this.isLinked && !this.isVertical;
+    },
+    leftButtonHidden() {
+      return this.position === 2 && this.isLinked && !this.isVertical;
+    },
+    maxPage_() {
+      return (
+        this.maxPage || (this.maxDate && getPageForDate(this.maxDate)) || null
+      );
     },
     maxFromPage() {
-      if (this.isDoublePaned_) return getPrevPage(this.maxPage);
-      return this.maxPage;
+      if (this.isDoublePaned_) return getPrevPage(this.maxPage_);
+      return this.maxPage_;
     },
     minToPage() {
-      if (this.isDoublePaned_) return getNextPage(this.minPage);
+      if (this.isDoublePaned_) return getNextPage(this.minPage_);
       return null;
     },
     themeStyles_() {
@@ -96,11 +137,12 @@ export default {
       };
     },
     wrapperStyle() {
-      const minWidth = `${this.isDoublePaned_ ? this.paneWidth * 2 : this.paneWidth}px`;
-      return {
-        ...this.themeStyles_.wrapper,
-        minWidth,
-      };
+      return this.themeStyles_.wrapper;
+    },
+    dividerStyle() {
+      return this.isVertical
+        ? this.themeStyles_.horizontalDivider
+        : this.themeStyles_.verticalDivider;
     },
     attributes_() {
       return AttributeStore(this.attributes);
@@ -113,68 +155,81 @@ export default {
     toPage() {
       this.refreshToPage();
     },
-    fromPage_(val) {
+    fromPage_(val, oldVal) {
+      if (pageIsEqualToPage(val, oldVal)) return;
       this.$emit('update:frompage', val);
       this.$emit('update:fromPage', val);
-      if (!pageIsBeforePage(val, this.toPage_)) {
+      if (!this.isDoublePaned) return;
+      if (this.isLinked || !pageIsBeforePage(val, this.toPage_))
         this.toPage_ = getNextPage(val);
-      }
     },
-    toPage_(val) {
+    toPage_(val, oldVal) {
+      if (pageIsEqualToPage(val, oldVal)) return;
       this.$emit('update:topage', val);
       this.$emit('update:toPage', val);
-      if (this.isDoublePaned_ && !pageIsAfterPage(val, this.fromPage_)) {
+      if (!this.isDoublePaned) return;
+      if (this.isLinked || !pageIsAfterPage(val, this.fromPage_))
         this.fromPage_ = getPrevPage(val);
-      }
     },
     isDoublePaned_() {
+      this.refreshIsConstrained();
       this.refreshToPage();
+    },
+    isLinked(val) {
+      if (val) this.toPage_ = getNextPage(this.fromPage_);
+    },
+    isExpanded() {
+      this.refreshIsConstrained();
     },
   },
   created() {
-    this.handleResize();
-    window.addEventListener('resize', this.handleResize);
     this.refreshFromPage();
     this.refreshToPage();
   },
+  mounted() {
+    this.$nextTick(() => {
+      this.refreshIsConstrained();
+      window.addEventListener('resize', this.refreshIsConstrained);
+    });
+  },
   beforeDestroy() {
-    window.removeEventListener('resize', this.handleResize);
+    window.removeEventListener('resize', this.refreshIsConstrained);
   },
   methods: {
     refreshFromPage() {
       this.fromPage_ = getFirstValidPage(
-        ...[this.fromPage, { month: todayComps.month, year: todayComps.year }].map(p =>
-          getPageBetweenPages(p, this.minPage, this.maxPage),
-        ),
-        this.minPage,
-        getPrevPage(this.maxPage),
+        ...[
+          this.fromPage,
+          { month: todayComps.month, year: todayComps.year },
+        ].map(p => getPageBetweenPages(p, this.minPage_, this.maxPage_)),
+        this.minPage_,
+        getPrevPage(this.maxPage_),
       );
     },
     refreshToPage() {
       this.toPage_ = getFirstValidPage(
         ...[this.toPage, getNextPage(this.fromPage_)].map(p =>
-          getPageBetweenPages(p, this.minPage, this.maxPage),
+          getPageBetweenPages(p, this.minPage_, this.maxPage_),
         ),
-        this.maxPage,
-        getNextPage(this.minPage),
+        this.maxPage_,
+        getNextPage(this.minPage_),
       );
     },
-    titleClick(page) {
-      if (pageIsEqualToPage(page, todayComps)) return;
-      if (page.position < 2) {
-        this.fromPage_ = getFirstValidPage(
-          todayComps,
-          pageIsBeforePage(page, todayComps) ? this.maxFromPage : this.minPage,
-        );
+    refreshIsConstrained() {
+      // Get the root calendar element
+      const root = this.$refs.root;
+      // Only test for constrained environment if needed
+      if (!window || !root || !this.isDoublePaned || this.isVertical) {
+        this.isConstrained = false;
+        // Test for constrained window
+      } else if (window && window.innerWidth < 2 * this.paneWidth + 30) {
+        this.isConstrained = true;
+      } else if (this.isExpanded) {
+        this.isConstrained =
+          root.parentElement.offsetWidth < 2 * this.paneWidth + 2;
       } else {
-        this.toPage_ = getFirstValidPage(
-          todayComps,
-          pageIsBeforePage(page, todayComps) ? this.maxPage : this.minToPage,
-        );
+        this.isConstrained = false;
       }
-    },
-    handleResize() {
-      this.windowWidth = window.innerWidth;
     },
   },
 };
@@ -191,13 +246,16 @@ export default {
   font-weight: $font-weight
   line-height: 1.5
   color: $font-color
-  min-width: $pane-width
   -webkit-font-smoothing: antialiased
   -moz-osx-font-smoothing: grayscale
-  &.is-double-paned
-    min-width: $pane-width * 2
+  box-sizing: border-box
   &.is-expanded
     width: 100%
+  &.is-vertical
+    flex-direction: column
+  /deep/ *
+    box-sizing: inherit
+
 
 .c-pane-divider
   width: 1px