+++ src/utils/defaults.js
@@ -1,6 +1,5 @@
 import { POPOVER_VISIBILITIES } from './constants';
 import { isObject, isFunction } from './typeCheckers';
-import setupLocale from './locales';
 
 const defaults = {
   componentPrefix: 'v',
@@ -10,6 +9,7 @@ const defaults = {
   titleTransition: 'slide-h',
   weeksTransition: 'slide-h',
   paneWidth: 256, // px,
+  showLinkedButtons: false,
   formats: {
     title: 'MMMM YYYY',
     weekdays: 'W',
@@ -99,17 +99,15 @@ const defaults = {
   themeStyles: {
     wrapper: { backgroundColor: '#fafafa', border: '1px solid #dadada' },
     verticalDivider: { borderLeft: '1px solid #dadada' },
+    horizontalDivider: { borderTop: '1px solid #dadada' },
     pane: null,
     header: null,
     headerTitle: null,
     headerArrows: null,
-    headerVerticalDivider: null,
     headerHorizontalDivider: null,
     weekdays: null,
-    weekdaysVerticalDivider: null,
     weekdaysHorizontalDivider: null,
     weeks: null,
-    weeksVerticalDivider: null,
     dayCell: null,
     dayCellNotInMonth: { opacity: 0.4 },
     dayContent: null,
@@ -121,19 +119,18 @@ const defaults = {
       fontSize: '.8rem',
       whiteSpace: 'nowrap',
     },
+    navHeader: null,
+    navHeaderTitle: null,
+    navHeaderArrows: null,
+    navMonthCell: null,
+    navYearCell: null,
   },
 };
-// Uncomment this line when running unit tests
-if (process.env.NODE_ENV === 'test') setupLocale(null, defaults);
 
 export default defaults;
 
 export const resolveDefault = (def, args) =>
   (isObject(def) && def) || (isFunction(def) && def(args)) || def;
 
-export const mergeDefaults = otherDefaults => {
-  // Setup locale defaults if needed
-  setupLocale(otherDefaults && otherDefaults.locale, defaults);
-  // Assign the defaults
-  return Object.assign(defaults, otherDefaults);
-};
+export const mergeDefaults = (...defaultArgs) =>
+  Object.assign(defaults, ...defaultArgs);